Split integration tests into two types
The neow3j are now separated into read-only and write tests.
This allows to execute the read-only tests faster because they
can all reuse the same container.
